Officially renamed all necessary files
